Fix VL53L1X implementation.

+++ src/main/java/org/rivierarobotics/i2c/api/Vl53lx.java
@@ -54,7 +54,7 @@ public interface Vl53lx {
 
     void stopContinuous();
 
-    short read();
+    int read();
 
     boolean dataReady();
